Move all two-thumb typing options into their own settings screen
User feedback: the ten new options were buried inside Gesture Typing
under a 'Two-thumb typing' section, mixed with the existing gesture
sliders/toggles. Hard to find. This commit lifts them into a dedicated
top-level entry 'Two-thumb typing (experimental)' on the main settings
screen, with a clearer subsection layout inside.

New screen: TwoThumbTypingScreen.kt
* Reused the same SearchSettingsScreen + Setting() pattern as the other
  screens.
* Four subsections by user-facing behaviour rather than implementation:
  - Spacing & word commit (#1.1 manual spacing + sub-option + alternative
    #1.2 autospace grace; mutually exclusive)
  - Tap + swipe interactions (#1.3 tap-during-swipe + window, #1.4 tap-
    then-swipe promotion)
  - Layout extras (#2.3 apostrophe key toggle — layout edit still owed
    by user)
  - Recognition & debug (#2.1 dual-thumb hinting + midline, debug overlay)
* Short-circuits to an empty screen when JNI lib is missing or gesture
  typing is off, since every toggle is a no-op until then.

Wiring:
* SettingsNavHost: new 'two_thumb_typing' route + composable.
* MainSettingsScreen: new Preference row right under Gesture Typing in
  the Typing card, with the 'experimental' label and a one-line summary
  so the entry is self-explanatory.
* SettingsContainer.createSettings: also include createTwoThumbTypingSettings
  so the entries remain globally searchable.
* GestureTypingScreen: removed the two-thumb section from the items
  list AND the Setting() factory entries (they now live in
  createTwoThumbTypingSettings).
* strings.xml: new settings_screen_two_thumb_typing + summary, plus four
  subsection headers (spacing/taps/layout/hinting).

Verified clean with :app:compileStandardDebugJavaWithJavac,
:app:compileStandardDebugKotlin, and a full :app:assembleStandardDebug.
