+++ app/src/main/java/helium314/keyboard/settings/SettingsContainer.kt
@@ -14,6 +14,7 @@ import helium314.keyboard.settings.screens.createGestureTypingSettings
 import helium314.keyboard.settings.screens.createLayoutSettings
 import helium314.keyboard.settings.screens.createPreferencesSettings
 import helium314.keyboard.settings.screens.createToolbarSettings
+import helium314.keyboard.settings.screens.createTwoThumbTypingSettings
 
 class SettingsContainer(context: Context) {
     private val list = createSettings(context)
@@ -76,7 +77,8 @@ class Setting(
 private fun createSettings(context: Context) = createAboutSettings(context) + createAppearanceSettings(context) +
         createCorrectionSettings(context) + createPreferencesSettings(context) + createToolbarSettings(context) +
         createLayoutSettings(context) + createAdvancedSettings(context) +
-        createGestureTypingSettings(context) // Always include (options disabled until library loaded)
+        createGestureTypingSettings(context) + // Always include (options disabled until library loaded)
+        createTwoThumbTypingSettings(context)  // Two-thumb typing (experimental) — own screen
 
 object SettingsWithoutKey {
     const val EDIT_PERSONAL_DICTIONARY = "edit_personal_dictionary"
